Make the OData Extensions public
There are situations where this code is useful, for example when having to override some of the built-in behavior to do custom batching. Being able to access extensions like `GetClrType()` is incredibly useful, and making it public would minimize the possibility of duplicate code across codebases.

Also, the team should consider refactoring some of these helpers to live back in the foundational OData classes.